MAINT: refactor and comments
