+++ data-prepper-plugins/peer-forwarder/src/main/java/com/amazon/dataprepper/plugins/processor/peerforwarder/PeerForwarder.java
@@ -12,6 +12,10 @@ import io.opentelemetry.proto.trace.v1.ResourceSpans;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.net.InetAddress;
+import java.net.NetworkInterface;
+import java.net.SocketException;
+import java.net.UnknownHostException;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -33,10 +37,28 @@ public class PeerForwarder implements Processor<Record<ExportTraceServiceRequest
 
     private final Map<String, TraceServiceGrpc.TraceServiceBlockingStub> peerClients;
 
+    public static boolean isAddressDefinedLocally(final String address) {
+        final InetAddress inetAddress;
+        try {
+            inetAddress = InetAddress.getByName(address);
+        } catch (UnknownHostException e) {
+            return false;
+        }
+        if (inetAddress.isAnyLocalAddress() || inetAddress.isLoopbackAddress()) {
+            return true;
+        } else {
+            try {
+                return NetworkInterface.getByInetAddress(inetAddress) != null;
+            } catch (SocketException e) {
+                return false;
+            }
+        }
+    }
+
     public PeerForwarder(final PluginSetting pluginSetting) {
         peerForwarderConfig = PeerForwarderConfig.buildConfig(pluginSetting);
         dataPrepperIps = new ArrayList<>(new HashSet<>(peerForwarderConfig.getDataPrepperIps()));
-        peerClients = dataPrepperIps.stream().filter(ip -> !PeerForwarderUtils.isAddressDefinedLocally(ip))
+        peerClients = dataPrepperIps.stream().filter(ip -> !isAddressDefinedLocally(ip))
                 .collect(Collectors.toMap(ip-> ip, ip-> createGRPCClient(ip)));
         Collections.sort(dataPrepperIps);
     }
@@ -54,6 +76,8 @@ public class PeerForwarder implements Processor<Record<ExportTraceServiceRequest
         for (final String dataPrepperIp: dataPrepperIps) {
             groupedRS.put(dataPrepperIp, new ArrayList<>());
         }
+
+        // Group ResourceSpans by consistent hashing of traceId
         for (final Record<ExportTraceServiceRequest> record: records) {
             for (final ResourceSpans rs: record.getData().getResourceSpansList()) {
                 final List<Map.Entry<String, ResourceSpans>> rsBatch = PeerForwarderUtils.splitByTrace(rs);
@@ -66,16 +90,18 @@ public class PeerForwarder implements Processor<Record<ExportTraceServiceRequest
             }
         }
 
-        // Send ExportTraceServiceRequest to designated peer
+        // Buffer of requests to be exported to the downstream of the local data-prepper
         final List<Record<ExportTraceServiceRequest>> results = new ArrayList<>();
         for (final String dataPrepperIp: dataPrepperIps) {
             final TraceServiceGrpc.TraceServiceBlockingStub client = peerClients.getOrDefault(dataPrepperIp, null);
+            // Create ExportTraceRequest for storing single batch of spans
             ExportTraceServiceRequest.Builder currRequestBuilder = ExportTraceServiceRequest.newBuilder();
             int currSpansCount = 0;
             for (final ResourceSpans rs: groupedRS.get(dataPrepperIp)) {
                 final int rsSize = PeerForwarderUtils.getResourceSpansSize(rs);
                 if (currSpansCount >= peerForwarderConfig.getMaxNumSpansPerRequest()) {
                     final ExportTraceServiceRequest currRequest = currRequestBuilder.build();
+                    // Send the batch request to designated remote peer or ingest into localhost
                     processRequest(client, currRequest, results);
                     currRequestBuilder = ExportTraceServiceRequest.newBuilder();
                     currSpansCount = 0;
@@ -83,7 +109,7 @@ public class PeerForwarder implements Processor<Record<ExportTraceServiceRequest
                 currRequestBuilder.addResourceSpans(rs);
                 currSpansCount += rsSize;
             }
-            // last request
+            // Dealing with the last batch request
             if (currSpansCount > 0) {
                 final ExportTraceServiceRequest currRequest = currRequestBuilder.build();
                 processRequest(client, currRequest, results);
